requirements.txt
python3
python3-waitress
python3-flask
python3-flask-login
python3-yaml
python3-passlib
